Module DkmlDuneDslShow.I

An interpreter of the Dune DSL whose interpreted result is a s-exp of a valid Dune file parameterizable by Mustache JSON configuration.

Use plain_hum or pretty to show the Dune file.

type 'a repr = Mustache.Json.t -> Sexplib.Sexp.With_layout.t

A type that represents a Dune file.

Stanzas

val rule : [ `RuleClause ] repr list -> [ `Stanza ] repr

The rule [rule_clause1 ...] stanza is used to create custom user rules. It tells Dune how to generate a specific set of files from a specific set of dependencies.

val executable : [ `Executable ] repr list -> [ `Stanza ] repr

val install : [ `Install ] repr list -> [ `Stanza ] repr

Rules

val alias : string -> [ `RuleClause ] repr

val targets : string list -> [ `RuleClause ] repr

val target : string -> [ `RuleClause ] repr

val deps : [ `Dep ] repr list -> [ `RuleClause ] repr

val action : [ `Action ] repr -> [ `RuleClause ] repr

action <action> is what you run to produce the targets from the dependencies.

Dependencies

val glob_files : string -> [ `Dep ] repr

val named_dep : name:string -> string -> [ `Dep ] repr

Actions

val echo : string list -> [ `Action ] repr

val with_stdout_to : string -> [ `Action ] repr -> [ `Action ] repr

val progn : [ `Action ] repr list -> [ `Action ] repr

val run : string list -> [ `Action ] repr

val diff : actual:string -> expected:string -> [ `Action ] repr

val diff_q : actual:string -> expected:string -> [ `Action ] repr

Executables

val public_name : string -> [ `Executable ] repr

val name : string -> [ `Executable ] repr

val libraries : string list -> [ `Executable ] repr

val modules : string list -> [ `Executable ] repr

val modes_byte_exe : [ `Executable ] repr

val ocamlopt_flags : [ `OCamlOptFlag ] repr list -> [ `Executable ] repr

Install

val section : string -> [ `Install ] repr

val install_files : [ `InstallDestination ] repr list -> [ `Install ] repr

https://dune.readthedocs.io/en/stable/dune-files.html#install-1.

Can either be:

      (files   (mylib.el as emacs/site-lisp/mylib.el))

or

      (files hello.txt)

Install Destination

val destination_file : filename:string -> destination:string -> [ `InstallDestination ] repr

https://dune.readthedocs.io/en/stable/dune-files.html#install-1.

Represents:

      (mylib.el as emacs/site-lisp/mylib.el)

in

        (files   (mylib.el as emacs/site-lisp/mylib.el))